dataset_analysis.py
"""
Dataset Analysis Script
======================

This script analyzes the dataset folder and demonstrates how the ice detection model
would work with your actual data.
"""

import pandas as pd
import numpy as np
import os
from PIL import Image
import matplotlib.pyplot as plt

def analyze_dataset():
    """Analyze the dataset folder structure and contents"""
    print("Mars Water-Ice Detection Dataset Analysis")
    print("=" * 45)
    
    # Analyze CSV files
    csv_dir = "dataset/csv"
    csv_files = os.listdir(csv_dir)
    print(f"Found {len(csv_files)} CSV files in dataset/csv:")
    for file in csv_files[:5]:  # Show first 5 files
        print(f"  - {file}")
    if len(csv_files) > 5:
        print(f"  ... and {len(csv_files) - 5} more files")
    
    # Analyze the main comprehensive dataset
    main_csv = "dataset/csv/comprehensive_ice_detection_datasets.csv"
    if os.path.exists(main_csv):
        df = pd.read_csv(main_csv)
        print(f"\nMain dataset contains {len(df)} entries")
        print("Columns in the dataset:")
        for col in df.columns:
            print(f"  - {col}")
        
        print("\nSample entries:")
        print(df[['Dataset Name', 'Planet/Mission', 'Data Type']].head())
    
    # Analyze image files
    image_dir = "dataset/images"
    image_files = [f for f in os.listdir(image_dir) if f.endswith('.png')]
    print(f"\nFound {len(image_files)} image files in dataset/images:")
    for file in image_files:
        file_path = os.path.join(image_dir, file)
        size = os.path.getsize(file_path) / (1024 * 1024)  # Size in MB
        print(f"  - {file} ({size:.1f} MB)")
    
    # Show sample images
    print("\nAnalyzing sample images...")
    fig, axes = plt.subplots(1, min(3, len(image_files)), figsize=(15, 5))
    if len(image_files) == 1:
        axes = [axes]
    
    for i, image_file in enumerate(image_files[:3]):
        img_path = os.path.join(image_dir, image_file)
        img = Image.open(img_path)
        if len(image_files) > 1:
            axes[i].imshow(img)
            axes[i].set_title(f"{image_file}\n{img.size[0]}x{img.size[1]} pixels")
            axes[i].axis('off')
        else:
            axes.imshow(img)
            axes.set_title(f"{image_file}\n{img.size[0]}x{img.size[1]} pixels")
            axes.axis('off')
    
    plt.tight_layout()
    plt.savefig('data/models/sample_images_analysis.png', dpi=300, bbox_inches='tight')
    plt.show()
    print("Sample images visualization saved as 'data/models/sample_images_analysis.png'")
    
    # Print dataset summary
    print("\nDataset Summary:")
    print("-" * 20)
    print(f"Total CSV files: {len(csv_files)}")
    print(f"Main dataset entries: {len(df) if 'df' in locals() else 'N/A'}")
    print(f"Image files: {len(image_files)}")
    print(f"Total size of images: {sum(os.path.getsize(os.path.join(image_dir, f)) for f in image_files) / (1024*1024):.1f} MB")

def demonstrate_model_architecture():
    """Demonstrate the model architecture that would be used"""
    print("\n\nIce Detection Model Architecture")
    print("=" * 35)
    
    model_description = """
    CNN Model for Mars Water-Ice Detection:
    
    Input: Satellite images (128x128x3)
    
    1. Convolutional Block 1:
       - Conv2D: 32 filters, 3x3 kernel
       - BatchNormalization
       - ReLU activation
       - MaxPooling2D: 2x2 pool size
    
    2. Convolutional Block 2:
       - Conv2D: 64 filters, 3x3 kernel
       - BatchNormalization
       - ReLU activation
       - MaxPooling2D: 2x2 pool size
    
    3. Convolutional Block 3:
       - Conv2D: 128 filters, 3x3 kernel
       - BatchNormalization
       - ReLU activation
       - MaxPooling2D: 2x2 pool size
    
    4. Global Average Pooling
    
    5. Dense Block:
       - Dense: 256 units
       - BatchNormalization
       - ReLU activation
       - Dropout: 0.5
    
    6. Output Layer:
       - Dense: 2 units (No Ice / Ice)
       - Softmax activation
    
    Total Parameters: ~1.2M
    """
    
    print(model_description)

def demonstrate_training_process():
    """Demonstrate how the training process would work"""
    print("\n\nTraining Process")
    print("=" * 15)
    
    training_description = """
    1. Data Preprocessing:
       - Load images and resize to 128x128
       - Normalize pixel values to [0, 1]
       - Extract labels from dataset metadata
       - Split data into train/validation/test sets (70%/15%/15%)
    
    2. Model Compilation:
       - Optimizer: Adam
       - Loss Function: Categorical Crossentropy
       - Metrics: Accuracy
    
    3. Training Configuration:
       - Epochs: 50 (with early stopping)
       - Batch Size: 32
       - Callbacks:
         * EarlyStopping (patience=10)
         * ReduceLROnPlateau (factor=0.2, patience=5)
    
    4. Evaluation:
       - Classification Report
       - Confusion Matrix
       - Accuracy/Loss curves
    """
    
    print(training_description)

if __name__ == "__main__":
    # Create output directory if it doesn't exist
    os.makedirs("data/models", exist_ok=True)
    
    # Run analysis
    analyze_dataset()
    demonstrate_model_architecture()
    demonstrate_training_process()
    
    print("\n\nAnalysis complete! The dataset trainer is ready to work with your data.")
    print("To run with actual model training, please install TensorFlow:")
    print("  pip install tensorflow")