requirements.txt
# LunarVision AI - Required Python Packages

# Image Processing
opencv-python>=4.5.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
tensorflow>=2.8.0
scikit-learn>=1.0.0
scikit-image>=0.18.0

# Data Visualization
matplotlib>=3.4.0
plotly>=5.0.0
seaborn>=0.11.0

# Web Framework
flask>=2.0.0
reportlab>=3.6.0

# Utilities
pandas>=1.3.0
pillow>=8.3.0